Added sub-grid cloud variablity to P3
This commit represents changes made to P3 to allow for sub-grid cloud
variability.  The cloud fraction is taken from macrophysics, in this
case CLUBB, and used to compute microphysics tendencies based on the
in cloud mass mixing ratios.  Sedimentation is still performed on the
cell-averaged quantities due to mass conservation requirements.  However
the fall-speeds of sedimentation are calculated based on the in-cloud
mass mixing ratios.